+++ demo-app/src/main/java/cash/z/ecc/android/sdk/demoapp/demos/getbalance/GetBalanceFragment.kt
@@ -14,7 +14,9 @@ import cash.z.ecc.android.sdk.demoapp.databinding.FragmentGetBalanceBinding
 import cash.z.ecc.android.sdk.demoapp.ext.requireApplicationContext
 import cash.z.ecc.android.sdk.demoapp.util.SyncBlockchainBenchmarkTrace
 import cash.z.ecc.android.sdk.demoapp.util.fromResources
+import cash.z.ecc.android.sdk.ext.convertZatoshiToZec
 import cash.z.ecc.android.sdk.ext.convertZatoshiToZecString
+import cash.z.ecc.android.sdk.ext.toUsdString
 import cash.z.ecc.android.sdk.internal.Twig
 import cash.z.ecc.android.sdk.model.Account
 import cash.z.ecc.android.sdk.model.PercentDecimal
@@ -23,9 +25,11 @@ import cash.z.ecc.android.sdk.model.Zatoshi
 import cash.z.ecc.android.sdk.model.ZcashNetwork
 import cash.z.ecc.android.sdk.tool.DerivationTool
 import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.launch
+import java.math.BigDecimal
 
 /**
  * Displays the available balance && total balance associated with the seed defined by the default config.
@@ -56,6 +60,16 @@ class GetBalanceFragment : BaseDemoFragment<FragmentGetBalanceBinding>() {
         val seedPhrase = sharedViewModel.seedPhrase.value
         val seed = Mnemonics.MnemonicCode(seedPhrase).toSeed()
         val network = ZcashNetwork.fromResources(requireApplicationContext())
+
+        binding.refreshExchangeRate.apply {
+            setOnClickListener {
+                lifecycleScope.launch {
+                    sharedViewModel.synchronizerFlow.value?.let { synchronizer ->
+                        synchronizer.refreshExchangeRateUsd()
+                    }
+                }
+            }
+        }
 
         binding.shield.apply {
             setOnClickListener {
@@ -100,38 +114,50 @@ class GetBalanceFragment : BaseDemoFragment<FragmentGetBalanceBinding>() {
                 launch {
                     sharedViewModel.synchronizerFlow
                         .filterNotNull()
-                        .flatMapLatest { it.saplingBalances }
+                        .flatMapLatest {
+                            it.saplingBalances.combine(it.exchangeRateUsd) { b, r ->
+                                b?.let { Pair(b, r?.first) }
+                            }
+                        }
                         .collect { onSaplingBalance(it) }
                 }
                 launch {
                     sharedViewModel.synchronizerFlow
                         .filterNotNull()
-                        .flatMapLatest { it.orchardBalances }
+                        .flatMapLatest {
+                            it.orchardBalances.combine(it.exchangeRateUsd) { b, r ->
+                                b?.let { Pair(b, r?.first) }
+                            }
+                        }
                         .collect { onOrchardBalance(it) }
                 }
                 launch {
                     sharedViewModel.synchronizerFlow
                         .filterNotNull()
-                        .flatMapLatest { it.transparentBalance }
+                        .flatMapLatest {
+                            it.transparentBalance.combine(it.exchangeRateUsd) { b, r ->
+                                b?.let { Pair(b, r?.first) }
+                            }
+                        }
                         .collect { onTransparentBalance(it) }
                 }
             }
         }
     }
 
-    private fun onOrchardBalance(orchardBalance: WalletBalance?) {
+    private fun onOrchardBalance(orchardBalance: Pair<WalletBalance, BigDecimal?>?) {
         binding.orchardBalance.apply {
-            text = orchardBalance.humanString()
+            text = orchardBalance.balanceHumanString()
         }
     }
 
-    private fun onSaplingBalance(saplingBalance: WalletBalance?) {
+    private fun onSaplingBalance(saplingBalance: Pair<WalletBalance, BigDecimal?>?) {
         binding.saplingBalance.apply {
-            text = saplingBalance.humanString()
+            text = saplingBalance.balanceHumanString()
         }
     }
 
-    private fun onTransparentBalance(transparentBalance: Zatoshi?) {
+    private fun onTransparentBalance(transparentBalance: Pair<Zatoshi, BigDecimal?>?) {
         binding.transparentBalance.apply {
             text = transparentBalance.humanString()
         }
@@ -140,7 +166,7 @@ class GetBalanceFragment : BaseDemoFragment<FragmentGetBalanceBinding>() {
             // This check is not entirely correct - it does not calculate the resulting fee with the new Proposal API
             // Note that the entire fragment-based old Demo app will be removed as part of [#973]
             visibility =
-                if ((transparentBalance ?: Zatoshi(0)).value > 0L) {
+                if ((transparentBalance?.first ?: Zatoshi(0)).value > 0L) {
                     View.VISIBLE
                 } else {
                     View.GONE
@@ -156,18 +182,21 @@ class GetBalanceFragment : BaseDemoFragment<FragmentGetBalanceBinding>() {
                 Synchronizer.Status.SYNCING -> {
                     SyncBlockchainBenchmarkTrace.Event.BLOCKCHAIN_SYNC_START
                 }
+
                 Synchronizer.Status.SYNCED -> {
                     SyncBlockchainBenchmarkTrace.Event.BLOCKCHAIN_SYNC_END
                 }
+
                 else -> null
             }
         traceEvents?.let { reportTraceEvent(it) }
 
         binding.textStatus.text = "Status: $status"
         sharedViewModel.synchronizerFlow.value?.let { synchronizer ->
-            onOrchardBalance(synchronizer.orchardBalances.value)
-            onSaplingBalance(synchronizer.saplingBalances.value)
-            onTransparentBalance(synchronizer.transparentBalance.value)
+            val rate = synchronizer.exchangeRateUsd.value?.first
+            onOrchardBalance(synchronizer.orchardBalances.value?.let { Pair(it, rate) })
+            onSaplingBalance(synchronizer.saplingBalances.value?.let { Pair(it, rate) })
+            onTransparentBalance(synchronizer.transparentBalance.value?.let { Pair(it, rate) })
         }
     }
 
@@ -180,23 +209,35 @@ class GetBalanceFragment : BaseDemoFragment<FragmentGetBalanceBinding>() {
 }
 
 @Suppress("MagicNumber")
-private fun WalletBalance?.humanString() =
+private fun Pair<WalletBalance, BigDecimal?>?.balanceHumanString() =
     if (null == this) {
         "Calculating balance"
     } else {
         """
-        Pending balance: ${pending.convertZatoshiToZecString(12)}
-        Available balance: ${available.convertZatoshiToZecString(12)}
-        Total balance: ${total.convertZatoshiToZecString(12)}
+        Pending balance: ${first.pending.convertZatoshiToZecString(12)} (${
+            second?.multiply(first.pending.convertZatoshiToZec())
+                .toUsdString()
+        } USD)
+        Available balance: ${first.available.convertZatoshiToZecString(12)} (${
+            second?.multiply(first.available.convertZatoshiToZec())
+                .toUsdString()
+        } USD)
+        Total balance: ${first.total.convertZatoshiToZecString(12)} (${
+            second?.multiply(first.total.convertZatoshiToZec())
+                .toUsdString()
+        } USD)
         """.trimIndent()
     }
 
 @Suppress("MagicNumber")
-private fun Zatoshi?.humanString() =
+private fun Pair<Zatoshi, BigDecimal?>?.humanString() =
     if (null == this) {
         "Calculating balance"
     } else {
         """
-        Balance: ${convertZatoshiToZecString(12)}
+        Balance: ${first.convertZatoshiToZecString(12)} (${
+            second?.multiply(first.convertZatoshiToZec())
+                .toUsdString()
+        } USD)
         """.trimIndent()
     }